+++ core/CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `stac::geoparquet::Compression`, even if geoparquet is not enabled ([#396](https://github.com/stac-utils/stac-rs/pull/396))
 - `Type` ([#397](https://github.com/stac-utils/stac-rs/pull/397))
 - `Collection::item_assets` and `ItemAsset` ([#404](https://github.com/stac-utils/stac-rs/pull/404))
+- A few extension methods on `Fields` ([#405](https://github.com/stac-utils/stac-rs/pull/405))
 
 ### Changed
 
@@ -26,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Removed
 
 - `Error::ReqwestNotEnabled` and `Error::GdalNotEnabled` ([#396](https://github.com/stac-utils/stac-rs/pull/382))
+- `Asset::extensions` ([#405](https://github.com/stac-utils/stac-rs/pull/405))
 
 ## [0.9.0] - 2024-09-05